Update submit_jobs.py
Boost library issue that prevented us from building on the FNAL-only nodes (only seen when using the LoadRawData tool) was resolved by creating a tmp folder and bind mounting that when we enter our container. The issue was arising due to the local tmp folder running out of space for the larger RawData files.